Moves precmd hook to base of plugin
For shell-specific hooks like precmd, we should be calling it in
init.zsh to call it out. For now, we're just calling vcs_info and
re-sourcing the ssh-agent and gpg-agent info (in case it is reloaded in
another tmux pane) but it could be more than that.

Note to self: I should really just create a hook system that allows me
to append shell functions to an overall chpwd and precmd meta function